Refine: better handling of failed DB saving (messages + Export option)
Squashed commits:
Refine: error message when write permission is denied
[682c16a1] Refactor(dbm): add support for showing saving error reason and recovery suggestion
[e932c951] Refactor: generalize DatabaseSavignViewController as DatabaseSaving protocol
[5f89d5fc] Feat: add possiblity to export database which could not be written
[23adee0a] Refine(files): a more informative error for OneDrive being read-only
[4b74192c] Refine(files): allow explicit cleanup of temporary files
[7ce510f7] Refactor(viewGroup): adopt Progress View Host
[7de83106] Refactor(editEntry): adopt ProgressViewHost

+++ KeePassiumLib/KeePassiumLib/files/TemporaryFileURL.swift
@@ -11,6 +11,8 @@ import Foundation
 public class TemporaryFileURL {
     public private(set) var url: URL
     
+    private var isErased = false
+    
     public init(fileName: String) throws {
         let fileManager = FileManager.default
         let tmpFileDir = fileManager.temporaryDirectory
@@ -31,9 +33,13 @@ public class TemporaryFileURL {
         cleanup()
     }
     
-    private func cleanup() {
+    public func cleanup() {
+        guard !isErased else {
+            return
+        }
         Diag.verbose("Will remove temporary file")
         try? FileManager.default.removeItem(at: url)
+        isErased = true
         Diag.debug("Temporary file removed")
     }
 }
